pyproject.toml
[project]
name = "feces-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Sami Moisio",email = "sami.moisio@elisa.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tk (>=0.1.0,<0.2.0)",
    "google-genai (>=1.2.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pygame (>=2.6.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
